+++ src/balancing_platform/joystick.py
@@ -6,7 +6,7 @@ from serial and returns them for further usage.
 
 Code by: Magnus Øye, Dated: 23.10-2018
 Contact: magnus.oye@gmail.com
-Website: https://github.com/magnusoy/
+Website: https://github.com/magnusoy/Balancing-Platform
 """
 
 # Importing package
@@ -21,6 +21,8 @@ class Joystick(object):
     def __init__(self):
         self.lastMove0 = 0
         self.lastMove1 = 0
+        self.rounded1 = 0
+        self.rounded2 = 0
         try:
             self.joystick = pygame.joystick.Joystick(0)  # create a joystick instance
             self.joystick.init()  # init instance
@@ -30,21 +32,19 @@ class Joystick(object):
 
     def getEvents(self):
         """docstring"""
-        rounded1 = 0
-        rounded2 = 0
         for event in pygame.event.get():
             if event.type == pygame.JOYAXISMOTION:  # Joystick
                 if self.joystick.get_axis(0) != self.lastMove0 or self.joystick.get_axis(1) != self.lastMove1:
                     self.lastMove0 = self.joystick.get_axis(0)
                     self.lastMove1 = self.joystick.get_axis(1)
-                    rounded1 = round(self.joystick.get_axis(0), 3)
-                    rounded2 = round(self.joystick.get_axis(1), 3)
-                    # (f'Axis 0: {rounded1}  |  Axis 1: {rounded2}')
+                    self.rounded1 = round(self.joystick.get_axis(0), 3)
+                    self.rounded2 = round(self.joystick.get_axis(1), 3)
+                    # print(f'Axis 0: {rounded1}  |  Axis 1: {rounded2}')
 
             if event.type == pygame.JOYBUTTONDOWN:
                 pass
                 # print("Joystick Button pressed")
-            return rounded1, rounded2
+        return self.rounded1, self.rounded2
 
 
 # Simple example of usage
@@ -52,4 +52,5 @@ if __name__ == '__main__':
     joystick = Joystick()
 
     while True:
-        joystick.getAxis()
+        coord = joystick.getEvents()
+        print(coord)