+++ src/balancing_platform/util/morphological_transformation.py
@@ -6,7 +6,7 @@ locating objects.
 
 Code by: Magnus Øye, Dated: 05.10-2018
 Contact: magnus.oye@gmail.com
-Website: https://github.com/magnusoy/
+Website: https://github.com/magnusoy/Balancing-Platform
 """
 
 # Importing packages